requirements.txt
# create a virtual environment named venv
# python3 -m venv venv

# activate the virtual environment
# source venv/bin/activate

# make sure that you are working in venv

# upgrade pip
# pip install --upgrade pip

# install the required packages
# pip install -r requirements.txt

jupyter==1.0.0
lazypredict==0.2.12  # IMPORTANT: from Supervised.py line 98 change 'sparse=False' to 'sparse_output=False'
matplotlib==3.9.0
pandas==2.2.2
numpy==2.0.0
scikit-learn==1.5.0

trimesh~=4.4.3  # pip install numpy trimesh numpy-stl
pyglet~=1.5.29  # pip install pyglet==2.0.16, pip install "pyglet<2"
manifold3d~=2.5.1

ttkbootstrap==1.10.1
pillow==10.4.0
cplex==22.1.1.2
numpy==2.0.1